# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Agendamento",
        "description": "API para sistema de agendamento",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/bookings/my-bookings": {
            "get": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Listar meus agendamentos",
                "description": "Retorna todos os agendamentos do usuário autenticado",
                "operationId": "8ec94cc909f472c4c594506eed1c3e78",
                "responses": {
                    "200": {
                        "description": "Lista de agendamentos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "bookings": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Booking"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bookings": {
            "post": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Criar agendamento",
                "description": "Cria um novo agendamento regular",
                "operationId": "e33fa5a6507581f9b1e0d710a9a130d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "location",
                                    "lesson_period",
                                    "start_time",
                                    "end_time",
                                    "is_evaluation_period"
                                ],
                                "properties": {
                                    "location": {
                                        "type": "string",
                                        "example": "Quadra 1"
                                    },
                                    "lesson_period": {
                                        "type": "integer",
                                        "maximum": 9,
                                        "minimum": 1,
                                        "example": 3
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2024-12-25 14:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2024-12-25 15:00:00"
                                    },
                                    "is_evaluation_period": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Aula de tênis"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Agendamento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Agendamento criado com sucesso!"
                                        },
                                        "booking": {
                                            "$ref": "#/components/schemas/Booking"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos ou horário não disponível"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bookings/friendly-match": {
            "post": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Criar amistoso",
                "description": "Cria um agendamento para amistoso (apenas 1ª e 3ª semana do mês)",
                "operationId": "f50f514abd76784640b3290ce8dc9660",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "location",
                                    "lesson_period",
                                    "start_time",
                                    "end_time",
                                    "is_evaluation_period"
                                ],
                                "properties": {
                                    "location": {
                                        "type": "string",
                                        "example": "Quadra 1"
                                    },
                                    "lesson_period": {
                                        "type": "integer",
                                        "maximum": 9,
                                        "minimum": 1,
                                        "example": 3
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2024-12-07 16:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2024-12-07 17:00:00"
                                    },
                                    "is_evaluation_period": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Amistoso de sábado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Amistoso agendado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Amistoso agendado com sucesso!"
                                        },
                                        "booking": {
                                            "$ref": "#/components/schemas/Booking"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Data inválida - amistosos só na 1ª e 3ª semana"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bookings/{booking}": {
            "delete": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Cancelar agendamento",
                "description": "Cancela um agendamento específico",
                "operationId": "69e442ca715abf5036d9e4c2b22af630",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "description": "ID do agendamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agendamento cancelado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Agendamento cancelado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão para cancelar este agendamento"
                    },
                    "400": {
                        "description": "Não é possível cancelar este agendamento"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Booking": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "location": {
                        "type": "string",
                        "example": "Quadra 1"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "string",
                        "example": "scheduled"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Aula de tênis"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}